Added a dataProvider.js for handling incoming data
